refactor: restructure folders per system
