+++ src/Constants.tsx
@@ -8,9 +8,4 @@ export class Constants {
 	 * in ms, movement per cell
 	 */
 	public static readonly ANIMATION_UNIT_SPEED = 250;
-
-	public static readonly LAYER_TILES = 0;
-	public static readonly LAYER_DEAD = 30;
-	public static readonly LAYER_UNITS = 50;
-	public static readonly LAYER_UI = 100;
 }
