--- src/components/DebugComponent.ts
@@ -1,3 +1,0 @@
-import { TagComponent } from 'ecsy';
-
-export class DebugComponent extends TagComponent { }
