--- src/components/FpsComponent.ts
@@ -1,3 +1,0 @@
-import { TagComponent } from 'ecsy';
-
-export class FpsComponent extends TagComponent {}
