--- src/components/LayerComponent.ts
@@ -1,13 +1,0 @@
-import { Component, Types } from 'ecsy';
-
-export interface ILayerComponentProps {
-	layer: number;
-}
-
-export class LayerComponent extends Component<ILayerComponentProps> {
-	declare layer: number;
-}
-
-LayerComponent.schema = {
-	layer: { type: Types.Number, default: 10 },
-};
