--- src/components/SelectorComponent.ts
@@ -1,3 +1,0 @@
-import { TagComponent } from 'ecsy';
-
-export class SelectorComponent extends TagComponent {}
