+++ src/systems/PathfindingComponent.ts
