+++ src/systems/alive/AliveComponent.ts
