+++ src/systems/alive/AliveSystem.ts
@@ -1,9 +1,8 @@
 import { System, SystemQueries, Entity } from 'ecsy';
-import { AliveComponent } from '../components/AliveComponent';
-import { HealthComponent } from '../components/HealthComponent';
-import { LayerComponent } from '../components/LayerComponent';
-import { SelectableComponent } from '../components/SelectableComponent';
-import { Constants } from '../Constants';
+
+import { AliveComponent } from './AliveComponent';
+import { HealthComponent } from '../health/HealthComponent';
+import { SelectableComponent } from '../selection/SelectableComponent';
 
 export class AliveSystem extends System {
 	public static queries: SystemQueries = {
@@ -37,9 +36,5 @@ export class AliveSystem extends System {
 	private handleDead(entity: Entity): void {
 		entity.removeComponent(SelectableComponent);
 		entity.removeComponent(HealthComponent);
-		const layerComponent = entity.getMutableComponent(LayerComponent);
-		if (layerComponent) {
-			layerComponent.layer = Constants.LAYER_DEAD;
-		}
 	}
 }
