+++ src/systems/health/HealthComponent.ts
