+++ src/systems/movement/MovePositionDirectComponent.ts
@@ -1,5 +1,6 @@
 import { Component, Types } from 'ecsy';
-import { Vector2 } from '../math/Vector2';
+
+import { Vector2 } from '../../math/Vector2';
 
 export interface IMovePositionDirectComponentProps {
 	movePosition: Vector2 | null;
