+++ src/systems/movement/MoveTransformVelocityComponent.ts
@@ -1,5 +1,5 @@
 import { Component, Types } from 'ecsy';
-import { Vector2 } from '../math/Vector2';
+import { Vector2 } from '../../math/Vector2';
 
 export interface IMoveTransformVelocityComponentProps {
 	moveSpeed?: number;
