+++ src/systems/movement/MoveVelocityComponent.ts
@@ -1,5 +1,6 @@
 import { Component, Types } from 'ecsy';
-import { Vector2 } from '../math/Vector2';
+
+import { Vector2 } from '../../math/Vector2';
 
 export interface IMoveVelocityComponentProps {
 	moveSpeed?: number;
