+++ src/systems/render/sprite/SpriteComponent.ts
